+++ src/models/session.rs
@@ -1,4 +1,4 @@
-use chrono::{DateTime, Utc};
+use chrono::{DateTime, Local};
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, sqlx::FromRow, Serialize, Deserialize)]
@@ -6,6 +6,7 @@ pub struct Session {
     pub id: Option<i32>,
     pub app_name: String,
     pub window_name: Option<String>,
-    pub start_time: DateTime<Utc>,
+    pub start_time: DateTime<Local>,
     pub duration: i64, // in seconds
+    pub category: Option<String>,
 }